api/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-multipart
python-jose
python-dotenv
bcrypt
sqlalchemy[asyncio]
typing
asyncpg